Много многопотока

+++ Source/App/ObjectEditor.cs
@@ -118,7 +118,8 @@ namespace Object_tool
 
 		// Settings
 		public bool USE_OLD_BONES = true;
-		public bool ALLOW_LOG = false;
+		public bool ALLOW_LOG = true;
+		public bool DISABLE_LOG = false;
 
 		public Object_Editor()
 		{
@@ -194,6 +195,8 @@ namespace Object_tool
 			NoCompress = true;
 			dbg_window = true;
 			showWindowToolStripMenuItem.Enabled = false;
+			ALLOW_LOG = false; // Not stable
+			DISABLE_LOG = true; // Not stable
 #endif
 
 			if (USE_OLD_BONES)
@@ -447,54 +450,60 @@ namespace Object_tool
 				EditorProcess.StartInfo.UseShellExecute = dbg_window;
 				EditorProcess.Start();
 
-				LogTextBox.Text = "";
 				string log = Application.ExecutablePath.Substring(0, Application.ExecutablePath.LastIndexOf('\\')) + "\\visual_log.log";
-				if (File.Exists(log))
-					File.Delete(log);
-
 				Thread LogThread = null;
-				if (ALLOW_LOG)
+				if (!DISABLE_LOG)
 				{
-					LogThread = new Thread(() =>
+					LogTextBox.Text = "";
+					if (File.Exists(log))
+						File.Delete(log);
+
+					if (ALLOW_LOG)
 					{
-						while (true)
+						LogThread = new Thread(() =>
 						{
-							if (File.Exists(log))
+							while (true)
 							{
-								LogTextBox.Text = File.ReadAllText(log);
-								LogTextBox.SelectionStart = LogTextBox.TextLength;
-								LogTextBox.ScrollToCaret();
+								if (File.Exists(log))
+								{
+									LogTextBox.Text = File.ReadAllText(log);
+									LogTextBox.SelectionStart = LogTextBox.TextLength;
+									LogTextBox.ScrollToCaret();
+								}
+								Thread.Sleep(100);
 							}
-							Thread.Sleep(100);
-						}
-					});
-					LogThread.Start();
+						});
+						LogThread.Start();
+					}
 				}
 
                 EditorProcess.WaitForExit();
 				EditorWorking = false;
 
-				if (ALLOW_LOG)
+				if (ALLOW_LOG && !DISABLE_LOG)
 					LogThread.Abort();
 
-				if (EditorProcess.ExitCode < -100 || EditorProcess.ExitCode > 100) // 100% Error
-                {
-					log = Application.ExecutablePath.Substring(0, Application.ExecutablePath.LastIndexOf('\\')) + "\\engine.log";
-					if (File.Exists(log))
+				if (!DISABLE_LOG)
+				{
+					if (EditorProcess.ExitCode < -100 || EditorProcess.ExitCode > 100) // 100% Error
 					{
-						LogTextBox.Text += "\n\nERROR LOG:\n\n" + File.ReadAllText(log);
+						log = Application.ExecutablePath.Substring(0, Application.ExecutablePath.LastIndexOf('\\')) + "\\engine.log";
+						if (File.Exists(log))
+						{
+							LogTextBox.Text += "\n\nERROR LOG:\n\n" + File.ReadAllText(log);
+							LogTextBox.Text += "\n" + GetTime();
+							LogTextBox.SelectionStart = LogTextBox.TextLength;
+							LogTextBox.ScrollToCaret();
+						}
+					}
+					else if (File.Exists(log))
+					{
+						LogTextBox.Text = File.ReadAllText(log);
 						LogTextBox.Text += "\n" + GetTime();
 						LogTextBox.SelectionStart = LogTextBox.TextLength;
 						LogTextBox.ScrollToCaret();
+						File.Delete(log);
 					}
-				}
-				else if (File.Exists(log))
-				{
-					LogTextBox.Text = File.ReadAllText(log);
-					LogTextBox.Text += "\n" + GetTime();
-					LogTextBox.SelectionStart = LogTextBox.TextLength;
-					LogTextBox.ScrollToCaret();
-					File.Delete(log);
 				}
 				return EditorProcess.ExitCode;
 			}
